Simplify plot scripts
